Refactor KMZGenerator class into a separate file
Refactor the `KMZGenerator` class into a separate file.

* Add `pyatmo/kmz_generator_class.py`:
  - Move the `KMZGenerator` class definition to this new file.
  - Import necessary modules and functions.
  - Define the `KMZGenerator` class and its methods.
  - Set up the logger for the `KMZGenerator` class.

* Modify `pyatmo/kmz_generator.py`:
  - Remove the `KMZGenerator` class definition.
  - Import the `KMZGenerator` class from `pyatmo/kmz_generator_class.py`.
  - Initialize and run the `KMZGenerator` class.

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/LuisTellezSirocco/GeoAtmoPlot?shareId=XXXX-XXXX-XXXX-XXXX).
